:recycle: Refactor lib structure and enhance file sending
This commit performs two significant actions:

Firstly, the structure of the `lib` directory is refactored. The generic index.ts is removed, focusing on a more specific `post.ts` file that contains functions for generating UUIDs and posting data.

Secondly, the `+page.svelte` is improved by introducing a function that allows users to send files. This enhancement also includes UI changes like a new "Send" button and adjustments in the layout for better UX.

The reason for these changes is twofold. Cleaning up the `lib` structure makes it easier for developers to locate and manage utility functions. Additionally, the sending functionality in `+page.svelte` provides users with more interactivity and utility, allowing for a smoother experience when working with files.